Finished the algorithm's implementation... kinda.
The algorithm terminates properly now, and computes the correct MST for all the
cases I tried with. The only thing is... the way we forced termination is not
technically specified in the algorithm's publication. The algorithm as it was
specified would not terminate, except for the two nodes who were the leader of
the final fragment. Therefore, we used a little hack where we broadcasted the
termination to all the other nodes before terminating the leaders themselves,
so everyone gets the memo and terminates as well.